+++ apps/desktop/package.json
@@ -57,15 +57,12 @@
     "electron-vite": "^3.0.0",
     "execa": "^9.5.2",
     "fix-path": "^4.0.0",
-    "http-proxy-agent": "^7.0.2",
-    "https-proxy-agent": "^7.0.6",
     "just-diff": "^6.0.2",
     "lodash": "^4.17.21",
     "lodash-es": "^4.17.21",
     "resolve": "^1.22.8",
     "semver": "^7.5.4",
     "set-cookie-parser": "^2.7.1",
-    "socks-proxy-agent": "^8.0.5",
     "tsx": "^4.19.3",
     "typescript": "^5.7.3",
     "undici": "^7.9.0",
