+++ src/locales/default/electron.ts
@@ -1,11 +1,14 @@
 const electron = {
   proxy: {
     auth: '需要认证',
+    authDesc: '如果代理服务器需要用户名和密码',
     authSettings: '认证设置',
     basicSettings: '代理设置',
+    basicSettingsDesc: '配置代理服务器的连接参数',
     bypass: '不使用代理的地址',
     connectionTest: '连接测试',
     enable: '启用代理',
+    enableDesc: '开启后将通过代理服务器访问网络',
     password: '密码',
     password_placeholder: '请输入密码',
     port: '端口',
@@ -17,8 +20,7 @@ const electron = {
     testButton: '测试连接',
     testDescription: '使用当前代理配置测试连接，验证配置是否正常工作',
     testFailed: '连接失败',
-    testSuccess: '连接成功',
-    testSuccessWithTime: '连接成功（{{time}}毫秒）',
+    testSuccessWithTime: '测试连接成功，耗时 {{time}} ms',
     testUrl: '测试地址',
     testUrlPlaceholder: '请输入要测试的 URL',
     testing: '正在测试连接...',
